requirements.txt
streamlit==1.50.0
numpy==1.24.3
pillow==11.3.0

tensorflow==2.13.0

protobuf==4.25.8
h5py==3.14.0
keras==2.13.1
typing_extensions==4.5.0
